requirements.txt
attrs==21.2.0
factory-boy==3.2.0
pytest==6.2.5
